chore: split agent data
