+++ build.config.ts
@@ -3,6 +3,7 @@ import { defineBuildConfig } from 'unbuild'
 export default defineBuildConfig({
   entries: [
     'src/index.ts',
+    'src/agents.ts',
   ],
   clean: true,
   declaration: true,
